Move password requirements to pydash_app.user
