+++ pydash/pydash_app/user/__init__.py
@@ -37,6 +37,9 @@ from . import repository
 from . import verification
 from multi_indexed_collection import DuplicateIndexError
 
+_MINIMUM_PASSWORD_LENGTH1 = 8
+_MINIMUM_PASSWORD_LENGTH2 = 12
+
 
 def add_to_repository(user):
     """
@@ -164,4 +167,18 @@ def verify(verification_code):
     verification.verify(verification_code)
 
 
+def check_password_requirements(password):
+    rules1 = [
+        lambda xs: any(x.isupper() for x in xs),
+        lambda xs: any(not x.isalpha() for x in xs),
+        lambda xs: len(xs) >= _MINIMUM_PASSWORD_LENGTH1
+    ]
+    rules2 = [
+        lambda xs: len(xs) >= _MINIMUM_PASSWORD_LENGTH2
+    ]
+    alternatives = [rules1, rules2]
 
+    def check_rules(rules):
+        return all(rule(password) for rule in rules)
+
+    return any(check_rules(alternative) for alternative in alternatives)
